requirements.txt
python-telegram-bot[job-queue]==21.3
supabase==2.5.1
python-dotenv==1.0.1
playwright>=1.40.0
httpx>=0.25.0
pytz>=2024.1
sentry-sdk>=1.40.0
